fix(dashboard): u64 deserialize overflow + capacity/deposit display formatting
on_chain_event_id in D1EventRow was typed Option<i64>, but Solana
on-chain event IDs are u64 values derived from hashes and can exceed
i64::MAX. The IslandDAO V4 demo event's ID (15159210065911600000)
overflowed i64, causing serde_json deserialization to fail in
get_event_raw(), which propagated as a 500 on GET /api/dashboard/live.
The event-resolution path is the only non-fault-tolerant part of the
handler, so a single bad column killed the whole endpoint.

Changed Option<i64> -> Option<u64> to match the type used everywhere
else in the codebase (domain EventConfig, frontend types, API
request/response structs). The write path (upsert_event) already
interpolated config.on_chain_event_id (u64) directly into SQL, so only
the read/deserialize path was broken.

Also fixed two display bugs in the dashboard header surfaced by the
same real production data:
  - Deposit showed raw atomic units (15000000) instead of human-readable
    USDC. Now uses the existing format_usdc() helper -> 15.00 USDC.
  - Capacity showed the deprecated legacy capacity column (0) instead of
    the per-format capacity model. EventDashboardMeta now carries
    in_person_capacity: Option<i64> (None/negative = unlimited, matching
    D1's -1 sentinel convention), rendered as 'Unlimited' when unset.

Deployed to production (version de82f4c5... -> latest).

+++ worker/src/db/events.rs
@@ -60,7 +60,11 @@ pub struct D1EventRow {
     pub promptpay_id: Option<String>,
     pub escrow_address: Option<String>,
     pub organizer_wallet: Option<String>,
-    pub on_chain_event_id: Option<i64>,
+    // u64 not i64 — Solana on-chain event IDs are derived from hashes and can
+    // exceed i64::MAX (e.g. 15159210065911600000 > 9223372036854775807).
+    // Typing as i64 caused serde_json deserialize failures on the live
+    // dashboard endpoint when an event's ID occupied the high u64 range.
+    pub on_chain_event_id: Option<u64>,
     pub refund_deadline_hours: Option<i64>,
     pub max_refundable_deposits: Option<i64>,
     pub description: Option<String>,
@@ -165,7 +169,7 @@ impl D1EventRow {
             escrow_address: self.escrow_address.clone().unwrap_or_default(),
             escrow_status,
             organizer_wallet: self.organizer_wallet.clone().unwrap_or_default(),
-            on_chain_event_id: self.on_chain_event_id.unwrap_or(0) as u64,
+            on_chain_event_id: self.on_chain_event_id.unwrap_or(0),
             refund_deadline_hours: self.refund_deadline_hours.unwrap_or(168) as u32,
             max_refundable_deposits: self.max_refundable_deposits.unwrap_or(0) as u32,
             description: self.description.clone().unwrap_or_default(),
